+++ src/main/java/com/nhnacademy/marketgg/client/config/LocaleConfig.java
@@ -1,0 +1,63 @@
+package com.nhnacademy.marketgg.client.config;
+
+import java.util.Locale;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.MessageSource;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.support.ResourceBundleMessageSource;
+import org.springframework.web.servlet.LocaleResolver;
+import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+import org.springframework.web.servlet.i18n.CookieLocaleResolver;
+import org.springframework.web.servlet.i18n.LocaleChangeInterceptor;
+
+/**
+ * 다국어 지원을 위한 설정 클래스입니다.
+ *
+ * @author 박세완
+ * @author 이제훈
+ */
+@Configuration
+public class LocaleConfig implements WebMvcConfigurer {
+
+    /**
+     * 사용자 언어 환경을 설정해주기 위한 Resolver 입니다.
+     *
+     * @return 사용자 언어환경을 설정해주고 초기값을 지닌 SessionLocalResolver 를 반환합니다.
+     * @since 1.0.0
+     */
+    @Bean
+    public LocaleResolver localeResolver() {
+        CookieLocaleResolver cookieLocaleResolver = new CookieLocaleResolver();
+        cookieLocaleResolver.setDefaultLocale(Locale.KOREAN);
+        cookieLocaleResolver.setCookieName("lang");
+
+        return cookieLocaleResolver;
+    }
+
+    @Bean
+    public LocaleChangeInterceptor localeChangeInterceptor() {
+        LocaleChangeInterceptor interceptor = new LocaleChangeInterceptor();
+        interceptor.setParamName("lang");
+
+        return interceptor;
+    }
+
+    @Bean
+    public MessageSource messageSource(@Value("${spring.messages.basename}") String basename,
+                                       @Value("${spring.messages.encoding}") String encoding) {
+
+        ResourceBundleMessageSource messageSource = new ResourceBundleMessageSource();
+        messageSource.setBasename(basename);
+        messageSource.setDefaultEncoding(encoding);
+
+        return messageSource;
+    }
+
+    @Override
+    public void addInterceptors(InterceptorRegistry registry) {
+        registry.addInterceptor(this.localeChangeInterceptor());
+    }
+
+}
